+++ apps/prairielearn/assets/scripts/instructorAssessmentGroupsClient.ts
@@ -1,0 +1,1 @@
+import './behaviors/custom-file-input.js';