+++ apps/prairielearn/elements/pl-rich-text-editor/pl-rich-text-editor.js
@@ -29,6 +29,23 @@
       ];
     }
 
+    options.modules.keyboard = {
+      bindings: {
+        tab: {
+          key: 9,
+          handler: () => {
+            // Overrides Quill tab behavior that inserts a tab character.
+            // Retains other Quill tab behaviours (indent list items or code
+            // blocks, switch cells in table), falling back to the browser's
+            // default behavior (focus on next focusable element) for
+            // accessibility.
+            // https://quilljs.com/docs/modules/keyboard#configuration
+            return true;
+          },
+        },
+      },
+    };
+
     let inputElement = $('#rte-input-' + uuid);
     let quill = new Quill('#rte-' + uuid, options);
     let renderer = null;
@@ -45,10 +62,31 @@
     if (contents && renderer) contents = renderer.makeHtml(contents);
     contents = rtePurify.sanitize(contents, rtePurifyConfig);
 
-    quill.setContents(quill.clipboard.convert(contents));
+    quill.setContents(quill.clipboard.convert({ html: contents }));
 
     quill.on('text-change', function () {
-      let contents = rtePurify.sanitize(quill.root.innerHTML, rtePurifyConfig);
+      // If a user types something and erases it, the editor will be blank, but
+      // the content will be something like `<p></p>`. In order to make sure
+      // this is treated as blank by the element's parse code and tagged as
+      // invalid if allow-blank is not true, we explicitly set the value to an
+      // empty string if the editor is blank. This is done using the `isBlank`
+      // method, used by Quill to determine if the placeholder should be shown.
+      // This is not a perfect solution, since it will not catch cases like
+      // content with only a space or a bulleted list without text, but we're ok
+      // with this caveat.
+      //
+      // An alternative solution would be to use the `getText` method, but this
+      // would cause a false positive for elements that are part of the answer
+      // but don't have a text (e.g., images).
+      //
+      // Because `isBlank` is undocumented, this solution may break in the
+      // future (see https://github.com/slab/quill/issues/4254). To ensure that
+      // the element continues to work if this method is removed, we use
+      // optional chaining in the call. This would cause the empty check to
+      // fail but not crash, so the element can continue working.
+      let contents = quill.editor?.isBlank?.()
+        ? ''
+        : rtePurify.sanitize(quill.getSemanticHTML(), rtePurifyConfig);
       if (contents && renderer) contents = renderer.makeMarkdown(contents);
       inputElement.val(
         btoa(
@@ -63,7 +101,7 @@
 
   // Override default implementation of 'formula'
 
-  var Embed = Quill.imports.parchment.Embed;
+  var Embed = Quill.import('blots/embed');
 
   class MathFormula extends Embed {
     static create(value) {