+++ apps/prairielearn/src/pages/userSettings/userSettings.ts
@@ -9,7 +9,7 @@ import { HttpStatusError } from '@prairielearn/error';
 import * as sqldb from '@prairielearn/postgres';
 
 import { getPurchasesForUser } from '../../ee/lib/billing/purchases.js';
-import { InstitutionSchema, ModeSchema, UserSchema } from '../../lib/db-types.js';
+import { InstitutionSchema, EnumModeSchema, UserSchema } from '../../lib/db-types.js';
 import { isEnterprise } from '../../lib/license.js';
 
 import { AccessTokenSchema, UserSettings } from './userSettings.html.js';
@@ -52,7 +52,7 @@ router.get(
     const { mode } = await sqldb.callRow(
       'ip_to_mode',
       [req.ip, res.locals.req_date, authn_user.user_id],
-      z.object({ mode: ModeSchema }),
+      z.object({ mode: EnumModeSchema }),
     );
 
     res.send(
@@ -77,7 +77,7 @@ router.post(
       const { mode } = await sqldb.callRow(
         'ip_to_mode',
         [req.ip, res.locals.req_date, res.locals.authn_user.user_id],
-        z.object({ mode: ModeSchema }),
+        z.object({ mode: EnumModeSchema }),
       );
       if (mode !== 'Public') {
         throw new HttpStatusError(403, 'Cannot generate access tokens in exam mode.');