+++ apps/prairielearn/src/tests/sprocs/check_assessment_access.test.ts
@@ -3,14 +3,14 @@ import { z } from 'zod';
 
 import * as sqldb from '@prairielearn/postgres';
 
-import { type Mode } from '../../lib/db-types.js';
+import { type EnumMode } from '../../lib/db-types.js';
 import * as helperDb from '../helperDb.js';
 
 const sql = sqldb.loadSqlEquiv(import.meta.url);
 
 async function checkAssessmentAccess(params: {
   assessment_id: string;
-  authz_mode: Mode;
+  authz_mode: EnumMode;
   authz_mode_reason?: string;
   course_role: string;
   course_instance_role: string;